+++ nisystemlink/clients/testmonitor/models/__init__.py
@@ -1,3 +1,17 @@
 from ._api_info import Operation, V2Operations, ApiInfo
+from ._result import Result
+from ._status import StatusType, Status
+from ._create_results_partial_success import CreateResultsPartialSuccess
+from ._update_results_partial_success import UpdateResultsPartialSuccess
+from ._delete_results_partial_success import DeleteResultsPartialSuccess
+from ._paged_results import PagedResults
+from ._create_result_request import CreateResultRequest
+from ._update_result_request import UpdateResultRequest
+from ._query_results_request import (
+    QueryResultsRequest,
+    QueryResultValuesRequest,
+    ResultField,
+    ResultProjection,
+)
 
 # flake8: noqa