+++ tests/integration/spec/test_spec.py
@@ -13,13 +13,16 @@ from nisystemlink.clients.spec.models import (
     CreateSpecificationsPartialSuccess,
     CreateSpecificationsRequest,
     CreateSpecificationsRequestObject,
+    CreateSpecificationsRequestObject,
     NumericConditionValue,
     QuerySpecificationsRequest,
     SpecificationLimit,
     SpecificationProjection,
+    SpecificationProjection,
     SpecificationType,
     StringConditionValue,
     UpdateSpecificationsRequest,
+    UpdateSpecificationsRequestObject,
     UpdateSpecificationsRequestObject,
 )
 from nisystemlink.clients.spec.utilities._dataframe_utilities import get_specs_dataframe